// src/API.ts
export type CreateTodoInput = {
  id?: string | null;
  sort?: number | null;
  name: string;
};
// ...
export type Todo = {
  __typename: 'Todo';
  id: string;
  sort?: number | null;
  name: string;
  createdAt: string;
  updatedAt: string;
};